Complete restructure of post handling, and other minor changes.
Posts and clones are now handled by `PostHandler`, via `GlobalVars`. Facilitating this was the creation of a 'ClonedPost' class.

Overall, this should fix a few outstanding bugs relating to tracking the flairs of wanted posts, and correctly tracking and grouping together cloned post. Additionally, a cloned entry will now persist across cycles, further aiding tracking at the slight cost of memory.

To balance this out, saving slightly on memory, `removed_posts` is now a local variable of `PostHandler.check_mod_log()`, instead of unnecessarily being an attribute of `GlobalVars`.

Although not yet implemented, the creation of `PostHandler` and `ClonedPost` paves the way for identifying the root post of a clone instance - that is, the post that Blossom chooses as the 'true' one. This can be identified as usually being the only clone with comments. In the future, this will allow us to automatically remove flaired clones, a functionality that, while in the program, has not so far been usable.

Some docstrings are also incomplete or absent within the new classes, and will be filled in in later commits.

Also removed duplicated bug fix from the changelog.

Signed-off-by: Murdo B. Maclachlan <murdomaclachlan@duck.com>